Move identity and google-oauth under src/modules.
Extract Google OAuth as its own module and group identity by controllers, routes, middleware, utils, services, and repositories.
