+++ src/modules/identity/middleware/optional-auth.middleware.ts
@@ -1,7 +1,7 @@
 import { Request } from "express";
 import jwt from "jsonwebtoken";
 
-import { IdentityConfig } from "./config";
+import { IdentityConfig } from "../config";
 
 export function makeOptionalAuthentication(config: IdentityConfig) {
   return (req: Request): boolean => {
