+++ src/modules/identity/repositories/account.repository.ts
@@ -1,4 +1,4 @@
-import { PrismaClient } from "../../generated/prisma/client";
+import { PrismaClient } from "../../../generated/prisma/client";
 
 export class AccountRepository {
   private readonly prisma: PrismaClient;
