+++ src/modules/identity/repositories/player.repository.ts
@@ -1,4 +1,4 @@
-import { PrismaClient } from "../../generated/prisma/client";
+import { PrismaClient } from "../../../generated/prisma/client";
 
 export class PlayerRepository {
   private readonly prisma: PrismaClient;
